Moving some modules to independent files
